added report ins-type-by-treatment, plus improvements to report management (#12)
added report ins-type-by-treatment, plus improvements to report management

+++ src/main/groovy/com/kinespherept/model/report/CountAndPercentCell.groovy
@@ -4,6 +4,6 @@ import groovy.transform.Canonical
 
 @Canonical
 class CountAndPercentCell {
-    int count = 0
-    int percent = 0
+    double count = 0.0
+    double percent = 0.0
 }
